Detect changes on output attributes
If the attributes (description or sensitive) attached to an output changes, we should detect that the plan contains changes.

To do so, we must persist the description in the state file in order to compare it to subsequent run.

Also, as the description becomes available in the state file, we output it along with the value and sensitive on terraform output command.